cilium: ipsec, support rolling updates
Currently, rolling updates may get stuck due to a time period between
when some set of nodes have started with encryption enabled and another
set exist without encryption enabled. At this point these two sets of
nodes can only communicate from non-encrypted to encrypted set. The
set with encryption enabled will encrypt traffic that will in turn be
dropped by the set that has yet to enable encryption.

To resolve we make encryption a property of the endpoint id. Keeping
the key identifier with the endpoint to inform cilium which key
should be used during an upgrade. Because we use the mark space
to identify keys we limit the key space to two keys currently.

After this key secrets will need to be updated to include an id
field as follows,

  keys: "1 hmac(sha256) 0123456789abcdef0123456789abcdef cbc(aes) 0123456789abcdef0123456789abcdef"

Where '1' is the id here. IDs are enforced to be less than 16. This
is a bit arbitrary we could go as high as 256 without hitting mark
bit limits. However, 16 feels sufficient and we cant take bits back
later so start low and bump up if needed. The id '0' is a special id
and should not be used it represents the absence of keys in the
datapath.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>